feat : 장르 가져오기 타입 지정 및 API 파일 생성
