+++ src/app/mypage/page.tsx
@@ -11,6 +11,7 @@ import React, { useState, useEffect } from 'react';
 import { Button } from '@mui/material';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import EditIcon from '@mui/icons-material/Edit';
+import fetchGenres from '../api/edit_genre';
 
 const theme = createTheme({
   palette: {
@@ -24,33 +25,8 @@ const theme = createTheme({
     },
   },
 });
-// 선택한 장르와 선택된 장르를 변경하는 함수
-interface Genre {
-  genre_id: number;
-  genre_name: string;
-  state: boolean;
-}
 
 export default function Mypage() {
-  const [genres, setGenres] = useState<Genre[]>([]);
-  const [loading, setLoading] = useState<boolean>(true); // 변경: loading 상태 타입 명시
-  const [error, setError] = useState<any>(null); // 변경: error 상태 타입 명시
-  const fetchGenres = async () => {
-    try {
-      // GET 요청을 보냅니다.
-      const response: AxiosResponse<any> = await axios.get(
-        '/api/v1/users/preference',
-      );
-      // 응답 데이터에서 장르 정보를 추출합니다.
-      const genresData: Genre[] = response.data.data.genres;
-      // 장르 정보를 설정합니다.
-      setGenres(genresData);
-      setLoading(false); // 변경: 데이터 로딩이 완료되면 loading 상태 변경
-    } catch (err) {
-      setError(err); // 변경: 에러 처리 수정
-      setLoading(false);
-    }
-  };
   const recentSongs = [
     {
       title: 'title1',
@@ -134,6 +110,11 @@ export default function Mypage() {
     },
   ];
 
+  // useEffect(() => {
+  //   fetchGenres();
+  // }, [loading]);
+  const [genres, setGenres] = useState<Genre[]>([]);
+
   const handleGenreToggle = (genreId: number) => {
     // 장르를 토글하여 선택 상태를 변경
     const updatedGenres = genres.map((genre) =>
@@ -141,9 +122,6 @@ export default function Mypage() {
     );
     setGenres(updatedGenres);
   };
-
-  if (loading) return <div>Loading...</div>;
-  if (error) return <div>Error: {error.message}</div>;
 
   return (
     <ThemeProvider theme={theme}>
