+++ src/app/types/edit_genre.ts
@@ -1,0 +1,6 @@
+// 선택한 장르와 선택된 장르를 변경하는 함수
+export interface Genre {
+  genre_id: number;
+  genre_name: string;
+  state: boolean;
+}
